LPS-132795 Move util and constants to API module
