+++ modules/apps/google-places/google-places-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Google Places API
+Bundle-SymbolicName: com.liferay.google.places.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.google.places.constants,\
+	com.liferay.google.places.util
