+++ modules/apps/google-places/google-places-api/src/main/java/com/liferay/google/places/util/GooglePlacesUtil.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.google.places.web.internal.util;
+package com.liferay.google.places.util;
 
-import com.liferay.google.places.web.internal.constants.GooglePlacesWebKeys;
+import com.liferay.google.places.constants.GooglePlacesWebKeys;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
