+++ modules/apps/google-places/google-places-api/src/main/resources/com/liferay/google/places/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
