+++ modules/apps/google-places/google-places-web/src/main/resources/META-INF/resources/init.jsp
@@ -17,6 +17,6 @@
 <%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.google.places.web.internal.constants.GooglePlacesWebKeys" %><%@
+<%@ page import="com.liferay.google.places.constants.GooglePlacesWebKeys" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %>
